+++ Services/UserService.cs
@@ -1,52 +1,88 @@
 using System.Net;
-using System.Net.NetworkInformation;
+using efcore_intro;
 using Microsoft.EntityFrameworkCore;
 
 public class UserService(ApplicationDbContext dbContext) : IUserService
 {
     private readonly ApplicationDbContext context = dbContext;
 
-    public async Task<ApiResponse<string>> AddUserAsync(User user)
+    public async Task<ApiResponse<string>> AddUserAsync(CreateUserDTO dto)
     {
+        var user = new User
+        {
+            Username = dto.Username,
+            Age = dto.Age
+        };
+
         await context.Users.AddAsync(user);
-        var r = await context.SaveChangesAsync();
+        await context.SaveChangesAsync();
+
         return new ApiResponse<string>(HttpStatusCode.OK, "Added successfully!");
     }
 
     public async Task<ApiResponse<string>> DeleteUserAsync(int id)
     {
-        var user = await context.Users.FindAsync(id) ?? new();
+        var user = await context.Users.FindAsync(id);
+
+        if (user == null)
+        {
+            return new ApiResponse<string>(HttpStatusCode.NotFound, "User not found");
+        }
+
         context.Users.Remove(user);
         await context.SaveChangesAsync();
+
         return new ApiResponse<string>(HttpStatusCode.OK, "Deleted successfully!");
     }
 
-    public async Task<ApiResponse<User>> GetUserAsync(int id)
+    public async Task<ApiResponse<UserDTO>> GetUserAsync(int id)
     {
-        var user = await context.Users.FirstOrDefaultAsync<User>(x => x.Id == id);
-        return new ApiResponse<User>(HttpStatusCode.OK, "User with given id", user ?? new());
+        var user = await context.Users
+            .Where(x => x.Id == id)
+            .Select(x => new UserDTO
+            {
+                Id = x.Id,
+                Username = x.Username,
+                Age = x.Age
+            })
+            .FirstOrDefaultAsync();
+
+        if (user == null)
+        {
+            return new ApiResponse<UserDTO>(HttpStatusCode.NotFound, "User not found");
+        }
+
+        return new ApiResponse<UserDTO>(HttpStatusCode.OK, "User found", user);
     }
 
-    public async Task<ApiResponse<List<User>>> GetUsersAsync()
+    public async Task<ApiResponse<List<UserDTO>>> GetUsersAsync()
     {
-        var users = await context.Users.ToListAsync();
-        return new ApiResponse<List<User>>(HttpStatusCode.OK, "list of users", users);
+        var users = await context.Users
+            .Select(x => new UserDTO
+            {
+                Id = x.Id,
+                Username = x.Username,
+                Age = x.Age
+            })
+            .ToListAsync();
+
+        return new ApiResponse<List<UserDTO>>(HttpStatusCode.OK, "List of users", users);
     }
 
-    public async Task<ApiResponse<string>> UpdateUserAsync(User user)
+    public async Task<ApiResponse<string>> UpdateUserAsync(UpdateUserDTO dto)
     {
-        var u = await context.Users.FindAsync(user.Id);
-        if (u == null)
-        {
-            return new ApiResponse<string>(HttpStatusCode.NotFound, "user not found for update");
-        }
-        else
-        {
+        var user = await context.Users.FindAsync(dto.Id);
 
-            u.Username = user.Username;
-            u.Age = user.Age;
-            await context.SaveChangesAsync();
-            return new ApiResponse<string>(HttpStatusCode.OK, "updated successfully");
+        if (user == null)
+        {
+            return new ApiResponse<string>(HttpStatusCode.NotFound, "User not found for update");
         }
+
+        user.Username = dto.Username;
+        user.Age = dto.Age;
+
+        await context.SaveChangesAsync();
+
+        return new ApiResponse<string>(HttpStatusCode.OK, "Updated successfully");
     }
 }
